requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
beautifulsoup4
streamlit
langchain-groq
fastapi
uvicorn
langserve